refactor(nat): Move static NAT network function to dedicated file
For consistency with masquerading and port forwarding, move the code for
the network function and for the main logic for static NAT to a
dedicated nf.rs file, only leaving dependency inclusion and re-exports
in mod.rs.
